Fully silence generate_test_data when run with -v=0
This mainly affects the test suite, which passes in -v=0 but the fixture
loading message from loaddata was obscuring the output in the middle of
the tests.

We still fail to forward -v=2 to loaddata if somebody calls
generate_test_data with that. This does not seem worth rewriting the
self.silent checks or storing the verbosity value in another attribute.

Reviewed-By: Jukka Karvonen <jukka.karvonen@helsinki.fi>