Make LDAP result deduplication more efficient
The 'if res not in results' check is inefficient for large lists of
results, which are not too improbable with a large user database.
Use a temporary set instead, and convert it to a list after deduplication.

There's a complication: we can't put the LDAP results in a set as-is
because dicts are not hashable. They need to be serialized first.

As the contents are all hashable and we don't care about item ordering,
roundtripping the result dicts to frozensets and back works fine. More
complex data structures would require something like JSON.

Reviewed-By: Jukka Karvonen <jukka.karvonen@helsinki.fi>