Added requirements
Allow setting and testing requirements for permissions and roles.
Functions to get missing requirements for identity and membership.

When testing requirements for the single membership, allow grace
period before failing the test, if the membership is active.
Failure time is stored in the membership, so grace is calculated
from the time the first requirement fails.

Requirements are created with predefined type and value. Supported
requirement types are:

contract: require a signed contract of contract template type
specified in the requirement value. If level is given, require
at least that version of contract.

attribute: identity is required to have attribute of the name
"value". email_address and phone_number are supported values.
If level is given, require at least that verification level
for the attribute.

assurance: identity must have assurance_level of the level or
higher.

external: future type reservation, which always fails for now.

Reviewed-By: Niko Tyni <ntyni@cc.helsinki.fi>