Fixed linking identity to user
If identity was imported from the external system, identity with
identifier but without user was created. This incorrectly returned
an error.

If identity is correctly identified during login, create user if it
does not exist. This was incorrectly checked in
_authenticate_create_user.

Fixed test where Google auth with existing identifier was using
incorrect authentication type.